refactor(library): split handlers/library/mod.rs into cohesive submodules
library/mod.rs was 4741 lines (plus an already-split episodes.rs from
PR #43). Split the remainder along its natural seams:

- reconcile.rs (469) — series resolver helpers: resolve_tracked_series,
  resolve_series_context, maybe_reconcile_mal_entry, reconcile_all_
  fallback_entries, populate_series_cover_urls, maybe_hydrate_cumulative_
  offset, should_persist_detail_cache, ReconcileReport
- pages.rs (990) — index, needs_review_page, series_detail, plus
  build_episodes (373 lines on its own) and the relation-card builder
  (build_relation_groups + merge_relation_metadata + richness helpers)
- crud.rs (658) — simple CRUD endpoints: add/remove series, set_folder,
  set_monitoring, set_episode_monitoring, set_allow_upgrades, set_search_
  overrides, set_manual_override, list_folders, reconcile_fallbacks
- search.rs (2431) — anilist_search, api_series_detail, auto_search_*,
  interactive_search_*, search_batch_releases, grab_batch_result, grab_
  interactive_result, auto_expand_library_from_pack*, run_auto_search_
  targets*, emit_auto_search_terminal, batch_episode_numbers, inline tests

mod.rs shrinks to 242 lines: shared template structs, form types, and
the Episode/RelationGroup/RelationCard data types.

main.rs, sonarr_compat.rs, and radarr_compat.rs updated to use full
submodule paths (handlers::library::search::auto_search_series etc.)
for utoipa/__path_X resolution, same pattern as PR #43.

No behavior change. cargo test still 480 passes.
