+++ src/handlers/library/episodes.rs
@@ -18,10 +18,10 @@ use crate::models::{config, episode_tags, grabbed_torrents};
 use crate::services::{auto_search, logger, media};
 use crate::AppState;
 
-use super::{
-    build_episodes, resolve_series_context, resolve_tracked_series, run_auto_search_targets,
-    Episode, MarkEpisodeFailedForm,
-};
+use super::pages::build_episodes;
+use super::reconcile::{resolve_series_context, resolve_tracked_series};
+use super::search::run_auto_search_targets;
+use super::{Episode, MarkEpisodeFailedForm};
 
 #[utoipa::path(
     post,
