+++ src/handlers/sonarr_compat.rs
@@ -762,10 +762,10 @@ pub async fn add_series(
             let id = s.id;
             tokio::spawn(async move {
                 tokio::time::sleep(std::time::Duration::from_secs(2)).await;
-                let _ = super::library::auto_search_series(
+                let _ = super::library::search::auto_search_series(
                     axum::extract::State(state_clone),
                     axum::extract::Path(id),
-                    axum::extract::Query(super::library::AutoSearchQuery::default()),
+                    axum::extract::Query(super::library::search::AutoSearchQuery::default()),
                 ).await;
             });
         }
@@ -835,10 +835,10 @@ pub async fn execute_command(
         && let Some(series_id) = body.series_id {
             let state_clone = state.clone();
             tokio::spawn(async move {
-                let _ = super::library::auto_search_series(
+                let _ = super::library::search::auto_search_series(
                     axum::extract::State(state_clone),
                     axum::extract::Path(series_id),
-                    axum::extract::Query(super::library::AutoSearchQuery::default()),
+                    axum::extract::Query(super::library::search::AutoSearchQuery::default()),
                 ).await;
             });
         }
